+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/viewmodel/FapHubCategoryViewModel.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.faphub.category.impl.viewmodel
 
+import androidx.datastore.core.DataStore
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.Pager
@@ -7,16 +8,20 @@ import androidx.paging.PagingConfig
 import androidx.paging.cachedIn
 import com.flipperdevices.bridge.dao.api.FapHubHideItemApi
 import com.flipperdevices.core.pager.loadingPagingDataFlow
+import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.faphub.category.impl.api.CATEGORY_OPEN_PATH_KEY
 import com.flipperdevices.faphub.dao.api.FapNetworkApi
 import com.flipperdevices.faphub.dao.api.model.FapCategory
 import com.flipperdevices.faphub.dao.api.model.SortType
+import com.flipperdevices.faphub.dao.api.model.SortType.Companion.toSortType
 import com.flipperdevices.faphub.target.api.FlipperTargetProviderApi
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import tangle.inject.TangleParam
 import tangle.viewmodel.VMInject
@@ -26,12 +31,23 @@ class FapHubCategoryViewModel @VMInject constructor(
     @TangleParam(CATEGORY_OPEN_PATH_KEY)
     private val category: FapCategory,
     fapHubHideItemApi: FapHubHideItemApi,
-    targetProviderApi: FlipperTargetProviderApi
+    targetProviderApi: FlipperTargetProviderApi,
+    private val dataStoreSettings: DataStore<Settings>
 ) : ViewModel() {
-    private val sortTypeFlow = MutableStateFlow(SortType.UPDATE_AT_DESC)
+    private val sortState by lazy {
+        dataStoreSettings.data
+            .map { it.selectedCatalogSort.toSortType() }
+            .stateIn(
+                viewModelScope,
+                SharingStarted.Lazily,
+                initialValue = SortType.UPDATE_AT_DESC
+            )
+    }
+
+    fun getSortTypeFlow(): StateFlow<SortType> = sortState
 
     val faps = combine(
-        sortTypeFlow,
+        sortState,
         targetProviderApi.getFlipperTarget(),
         fapHubHideItemApi.getHiddenItems()
     ) { sortType, target, hiddenItems ->
@@ -45,13 +61,15 @@ class FapHubCategoryViewModel @VMInject constructor(
         }.flow
     }.flatMapLatest { it }.cachedIn(viewModelScope)
 
-    fun getSortTypeFlow(): StateFlow<SortType> = sortTypeFlow
-
     fun getCategoryName() = category.name
 
     fun onSelectSortType(sortType: SortType) {
         viewModelScope.launch(Dispatchers.Default) {
-            sortTypeFlow.emit(sortType)
+            dataStoreSettings.updateData {
+                it.toBuilder()
+                    .setSelectedCatalogSort(sortType.toSelectedSortType())
+                    .build()
+            }
         }
     }
 }