+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/api/SelfUpdaterSourceApi.kt
@@ -1,0 +1,11 @@
+package com.flipperdevices.selfupdater.api
+
+import com.flipperdevices.selfupdater.models.SelfUpdateResult
+
+interface SelfUpdaterSourceApi {
+    suspend fun checkUpdate(manual: Boolean): SelfUpdateResult
+
+    fun getInstallSourceName(): String
+
+    fun isSelfUpdateCanManualCheck(): Boolean
+}