+++ components/selfupdater/api/src/main/kotlin/com/flipperdevices/selfupdater/models/SelfUpdateResult.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.selfupdater.models
+
+enum class SelfUpdateResult {
+    COMPLETE,
+    NO_UPDATES,
+    IN_PROGRESS,
+    ERROR
+}