+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-# 1.6.4 - In progress
+# 1.6.5 - In progress
+
+- [Feature] Scroll on Infrared Screen
+- [Feature] Check Self Update App in Options (only for github)
+- [Feature] Fap Catalog save sort
+- [Feature] Check Self Update App in Optiond (only for github)
+- [FIX] Use by default dark theme in Wear OS
+- [FIX] Use default splashscreen in Wear OS
+- [FIX] Fix compose layout in Wear OS
+
+# 1.6.4 
 
 - [Feature] Add transparent flipper mockup
 - [Feature] Add broke rpc ble session button to debug settings
@@ -8,6 +18,7 @@
 - [Feature] Add fap manifest caching
 - [Feature] Not send empty report in faphub
 - [Feature] Add hide app button
+- [Feature] Check Self Update App in Optiond (only for github)
 - [CI] Add apk artifacts to Pull Requests
 - [CI] Fix failed build in merge_group
 - [CI] Bump deps except Sentry-gradle